refactoring edit and admin page
