+++ sage-android/Sage/app/src/main/res/values/strings_sign_up.xml
@@ -6,6 +6,7 @@
     <string name="sign_up_profile">Profile</string>
     <string name="sign_up_password">Password</string>
     <string name="sign_up_password_confirm">Confirm Password</string>
+    <string name="sign_up_current_password">Current Password</string>
     <string name="sign_up_first_name">First Name</string>
     <string name="sign_up_last_name">Last Name</string>
